backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
crewai
